// types/financial-statement/calk/bumdes-administrator.ts
export interface BumdesAdministratorFormData {
  nama_penasihat?: string;
  nama_ketua_operasional?: string;
  nama_sekretaris_operasional?: string;
  nama_bendahara_operasional?: string;
  nama_unit_operasional?: string;
  nama_kepala_unit_operasional?: string;
  nama_anggota_operasional?: string;
  nama_kepala_anggota_operasional?: string;
  nama_ketua_pengurus?: string;
  nama_sekretaris_pengurus?: string;
  nama_bendahara_pengurus?: string;
  nama_unit_pengurus?: string;
}